+++ vite.config.ts
@@ -10,8 +10,5 @@ export default defineConfig({
       formats: ["es", "cjs", "iife"],
       fileName: (format) => `hot-connect.${format}.js`,
     },
-    rollupOptions: {
-      external: ["@walletconnect/sign-client"],
-    },
   },
 });